chore(text-editor): move external files out of internal directory
