+++ src/components/text-editor/text-editor.component.js
@@ -21,17 +21,17 @@ import {
   hasInlineStyle,
   hasBlockStyle,
   blockStyleFn,
-} from "./utils";
+} from "./__internal__/utils";
 import {
   StyledEditorOutline,
   StyledEditorContainer,
 } from "./text-editor.style";
-import Counter from "./editor-counter";
-import Toolbar from "./toolbar";
-import Label from "../../../__experimental__/components/label";
-import Events from "../../../utils/helpers/events/events";
-import createGuid from "../../../utils/helpers/guid";
-import LabelWrapper from "./label-wrapper";
+import Counter from "./__internal__/editor-counter";
+import Toolbar from "./__internal__/toolbar";
+import Label from "../../__experimental__/components/label";
+import Events from "../../utils/helpers/events/events";
+import createGuid from "../../utils/helpers/guid";
+import LabelWrapper from "./__internal__/label-wrapper";
 
 const NUMBERS = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"];
 const INLINE_STYLES = ["BOLD", "ITALIC"];
@@ -363,8 +363,7 @@ TextEditor.propTypes = {
   rows: (props, propName) => {
     if (
       props[propName] &&
-      typeof props[propName] === "number" &&
-      props[propName] < 2
+      (typeof props[propName] !== "number" || props[propName] < 2)
     ) {
       return new Error(
         `Prop \`${propName}\` must be a number value greater than 2 to override the min-height of the \`TextEditor\``
