+++ src/components/text-editor/text-editor.d.ts
