+++ src/components/text-editor/text-editor.stories.mdx
@@ -1,16 +1,10 @@
 import { Meta, Props, Preview, Story } from "@storybook/addon-docs/blocks";
 import { useState, useRef } from "react";
-import EditorLink from "./editor-link";
 import TextEditor, {
   TextEditorState as EditorState,
   TextEditorContentState as ContentState,
 } from "./text-editor.component";
-import Toolbar from "./toolbar";
-import ToolbarButton from "./toolbar/toolbar-button";
-import EditorCounter from "./editor-counter";
-import IconButton from "../../icon-button";
-import Icon from "../../icon";
-import Button from "../../button";
+import Button from "../button";
 
 <Meta
   title="Design System/Text Editor"
@@ -240,8 +234,9 @@ Maybe try writing a little bit less?`
 </Preview>
 
 ### With custom row height
-THe `rows` prop allows for overriding the default min-height of the `TextEditor`. It accepts any number greater than 2 
+The `rows` prop allows for overriding the default min-height of the `TextEditor`. It accepts any number greater than 2 
 which is multiplied by the line-height (21px). 
+
 <Preview>
   <Story name="with custom row height">
     {() => {
